refactor: search 관련 로직 훅으로 분리
